Adds syndicate radio jammer.
Turn on by setting it to frequency 148.9
Prevents people sending or receiving radio messages within 6 tiles.
This range was established so a jammer left on the bridge next to the wall
will have the top AI intercom in range.
Affects headsets, intercoms and station bounced radios, robotic talk is unaffected.
Added var/global/jammers to keep track if one is jamming to avoid
unnecessary searches which might be costly since radio talk is common.