Adds blood clotting
Players will eventually stop bleeding on their own from minor injuries.
This is nowhere near as effective as a bandage, and takes roughly a
minute to close the smallest possible cut, depending on the RNG.  Heals
too slowly for large wounds, but you no longer have to go to the
hospital after a small cut if you plan on living for more than 20
minutes.