Adds SEKRIT gold slime functionality
Gold slimes will now produce either a borer or a swarmer when injected
with 5 units of strange reagent.  If/when they are rebalanced,
spiderlings and facehuggers can be added to that list as well.  The gold
slime will try to poplulate the creature from the current pool of
ghosts, and allow ghosts to enter the creature by clicking on it if
there are no ghosts availible at the time of creation.