Fixes ssair trying to proc unsimulated turfs.
An active/excited simulated turf would get deleted and converted to space, so this would change the turf in ssair's active_turfs list to a space turf, then it would runtime trying to call process_cell on it. (why byond!)

Before a : and some lack of caring and a type check handled this. I removed this thinking it wasn't needed when i did my lag fix. (My exact thought was "why would an unsimulated turf get in ssair's active turf list, that seems like a bug that shouldn't get hidden" i forgot that deleting a turf just converts it to a space turf and all existing references are updated by byond to point to the new space turf)

this OOP way seems better anyways.