Fix virus immunity and some tweaks.
The way diseases code works is you have curable, resistible and
carriable diseases (ONLY PREMADE DISEASES, HARDCODED, NOT MADE INGAME).
So right now, to have a disease removed from your body, i.e. cured, a
disease must have curable and resistible flags, that is the disease can
be cured and can be resisted after being cured (antibodies or whatever).
If a disease is only curable but not resistible, you can lower the
stage. but not remove it from your system, meaning cure it. Carriable is
useless and I don't know why it's in the code when it's not being used
anywhere.

So, all curable diseases are also necessarily resisted against because
of code (it shouldn't work that way IMO, diseases should probably
evolve, maybe mutadone, etc.). The list >>OF CHANGED/FIXED DISEASES<<
is:
CURABLE & RESISTIBLE:
-MAGNITIS
-GBS
-SPACE RETROVIRUS
-WIZARDITIS
-ALL TRANSFORMATION VIRUSES EXCEPT ZOMBIE

INCURABLE AND IRRESISTIBLE:
-TRANSFORMATION ZOMBIE (once infected, cannot lower stage and cannot
cure, you become a zombie)

Everything else is the same, curable and resistible, or whatever, they
haven't been touched. Imho code should be changed where curable means
actually curable and not "possibly lowering the stage but not actually
curing it" and some other bullshit. I'll do it, but I wanna know if
you're ok with how diseases are right now and if I should change any to
incurable, etc.